Added signalling packet. Changed database to save as much ram space as possible.

+++ VoiceCraft.Mobile/VoiceCraft.Mobile/Models/ServerModel.cs
@@ -4,8 +4,8 @@
     {
         public string Name { get; set; } = string.Empty;
         public string IP { get; set; } = string.Empty;
-        public int Port { get; set; } = 9050;
-        public uint Key { get; set; }
+        public ushort Port { get; set; } = 9050;
+        public ushort Key { get; set; }
         public int Codec { get; set; } = -1;
         public bool ClientSided { get; set; }
     }
